fix(root_search/newton): cast deltas to parameters.dtype (stabilize while_loop carry)
newton updater's deltas promoted to float64 while parameters stayed float32 ->
lax.while_loop carry dtype mismatch. Cast deltas to parameters.dtype.

black_scholes: 107 -> 113 passing.

+++ tf_quant_finance/math/root_search/newton.py
@@ -118,7 +118,7 @@ def root_finder(value_and_grad_func,
     def _updater(counter, parameters, converged, failed):
       """Updates each parameter via Newton's method."""
       values, gradients = value_and_grad_func(parameters)
-      deltas = tf.math.divide(values, gradients)
+      deltas = tf.cast(tf.math.divide(values, gradients), dtype=parameters.dtype)
 
       converged = tf.abs(
           deltas) < relative_tolerance * tf.abs(values) + tolerance
